Move structs into models module
