+++ apps/backend/src/api/types/versions.rs
@@ -1,4 +1,4 @@
-use crate::core::game::Game;
+use crate::models::game::Game;
 use serde::{Deserialize, Serialize};
 use ts_rs::TS;
