+++ apps/backend/src/core/mod.rs
@@ -1,6 +1,2 @@
 pub mod api_clients;
 pub mod bin_providers;
-pub mod files;
-pub mod game;
-pub mod secrets;
-pub mod server;
