+++ apps/backend/src/models/mod.rs
@@ -1,0 +1,4 @@
+pub mod files;
+pub mod game;
+pub mod secrets;
+pub mod server;
